Isolate shoals tide from levelgen rng
The builder calls some tide code when finishing up a shoals level,
roughly to initialize things. However, tide code is dependent on global
state, for things like shoals tide direction, and it is also randomized,
so it would have a different impact on the rng depending on this state.
This means that for incremental pregen (or just regular games before
incremental pregen) the outcome of later levelgen would be impacted by
the global tide state when entering a level. This commit ensures that
tide state only ever impacts the gameplay rng.